+++ scripts/macros/2024/classFeatures/ranger/gloomStalker/umbralSight.js
@@ -1,4 +1,4 @@
-import {umbralSight as umbralSightLegacy} from '../../../legacyMacros.js';
+import {umbralSight as umbralSightLegacy} from '../../../../../legacyMacros.js';
 export let umbralSight = {
     ...umbralSightLegacy,
     version: '1.3.81',